gcp_audit_buddy/pyproject.toml
[tool.poetry]
name = "gcp_audit_buddy_agent"
version = "0.1.0"
description = "GCP Auditing Assistant using Agent Development Kit"
authors = ["Sam Silverberg <sam.silverberg@gmail.com>"] # Please update this with your details
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
google-adk = "*"
google-cloud-aiplatform = "*"
google-genai = "*"
python-dotenv = "*"
google-cloud-bigquery = "*"
matplotlib = "*"
pandas = "*"
seaborn = "*"
svglib = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 